tested on desktop
need to see how much it has improved


Former-commit-id: f6e7ce0d44059173c53c3e085ef9c0d0042186b3

+++ core/v1/board.cpp
@@ -54,7 +54,7 @@ ComboList PBoard::eraseComboAndMoveOrbs(int *moveCount)
                 
                 // Start finding combos
                 floodfill(&combo, i, j, orb);
-                if (combo.size() >= minEraseCondition)
+                if ((int)combo.size() >= minEraseCondition)
                 {
                     moreCombo = true;
                     comboList.push_back(combo);
